Add a GdsApi::Search and deprecate ...::Rummager
